requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.5
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.5
colorama==0.4.6
discord-emoji==1.4.1
discord-webhook==1.1.0
discord.py==2.3.1
dnspython==2.4.0
docopt==0.6.2
exceptiongroup==1.1.2
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.3
multidict==6.0.4
n2w==0.1.3
priority==2.0.0
pymongo==4.4.1
python-dotenv==1.0.0
quart==0.18.4
redis==4.6.0
requests==2.31.0
sniffio==1.3.0
tomli==2.0.1
urllib3==2.0.3
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.2
